refactor MCP server and plugin
